Move legacy auction to endpoints package, update tests
